ISE 529 Predictive Analytics HW2
Fit a model and report the estimated regression coefficients (include a 99% confidence interval for each one.
Find the best and worst predictors in terms of correlation with predictors and R squared values.
Fit the model with the best subset of predictors (in terms of AIC). Report the fitted equation.
Fit a linear regression model and report the fitted equations.