ISE 529 Predictive Analytics HW 4
Question 1:
Fit a Random Forest model with B=100 bagged trees and max_features = 10. Report the test MSE. Determine the most important predictors.
Fit a Gradient boosting model with 100 trees and learning rate 0.10 and max_depth = 4. Report the test MSE.
Fit a multiple linear regression model with the most important predictors found by the Random Forest. Find the test MSE
Question 2:
Find the correlation matrix rXX for the predictors. Find the predictors which have the lowest correlation. Fit a linear regression of midarm circumference as a function of the other two predictors. Report the R2. Determine whether the predictors are independent or not.
Question 3:
Show there are two variables are highly unbalanced. And remove these variables from the dataset.
Fit a random forest model with 500 trees and max_features = 29 to the training set.
Fit a boosting model to the training set.
Report the test accuracy rate when KNN (with n_neighbors = 3) is used to predict the results.
Report the test accuracy rate when logistic regression is used to predict the results.